requirements.txt
torch
torchvision
tqdm
numpy
pillow
albumentations
opencv-python
torchmetrics